Move app-specific checks into a new appInfo.js file to avoid circular dependency issues.
