+++ kolibri/core/assets/src/core-app/apiSpec.js
@@ -52,6 +52,7 @@ import * as validators from '../validators';
 import * as serverClock from '../serverClock';
 import * as resources from '../api-resources';
 import * as i18n from '../utils/i18n';
+import * as appInfo from '../utils/appInfo';
 import * as browserInfo from '../utils/browserInfo';
 import { redirectBrowser } from '../utils/redirectBrowser';
 import * as licenseTranslations from '../utils/licenseTranslations';
@@ -191,6 +192,7 @@ export default {
   urls,
   utils: {
     appCapabilities,
+    appInfo,
     contentNode,
     redirectBrowser,
     browserInfo,
