+++ kolibri/core/assets/src/utils/browserInfo.js
@@ -5,8 +5,6 @@
  * the more general purpose code therein.
  */
 
-import store from 'kolibri.coreVue.vuex.store';
-
 const NAME = 'name';
 const VERSION = 'version';
 const browserTests = [
@@ -179,32 +177,6 @@ export function passesRequirements(browser, requirements) {
 
 export const userAgent =
   window && window.navigator && window.navigator.userAgent ? window.navigator.userAgent : '';
-
-/**
- * Detection of whether an Android device is using WebView based on
- * https://developer.chrome.com/multidevice/user-agent#webview_user_agent
- * First checks for 'wv' (Lolipop+), then for 'Version/x.x'
- */
-const isAndroid = /Android/.test(userAgent);
-const isMac = /Macintosh/.test(userAgent);
-const isWindows = /Windows/.test(userAgent);
-
-/**
- * All web views
- * TODO: Refactor calling code to check isAppContext explicitly rather than
- * calling this method. (Or ideally switch to calling checkAppCapabilities
- * for the functionality we want to expose, including file download.)
- *
- * Note that we explicitly exclude Linux from this check because the GNOME
- * app currently does not set this.
- */
-export const isEmbeddedWebView = (isAndroid || isMac || isWindows) && store.getters.isAppContext;
-
-/**
- * CoreFullscreen checks for embedded Android specifically
- * TODO: See if we can enable fullscreen on embedded Android
- */
-export const isAndroidWebView = isAndroid && isEmbeddedWebView;
 
 /**
  * General browser info
